add(processOrder method in Server Model) This commit initially implements the process order to the ServerModel class in the Server. It also includes the creation of  xml files for the purpose of saving customer account, order list and predefined food and beverages Further, this commit added the XMLUtility class where it is used to process xml reading and writing. However, the methods are not yet implemented. Lastly, this commit changed some bug in the Food Class

+++ src/main/java/shared/Food.java
@@ -15,8 +15,7 @@ public class Food extends Product {
 
     //getters setters
 
-
-    public double getQuantity() {
+    public int getQuantity() {
         return quantity;
     }
